+++ lang/en/theme_boost_union.php
@@ -46,3 +46,13 @@ $string['blocksgeneralheading'] = 'General blocks';
 
 // Privacy API.
 $string['privacy:metadata'] = 'The Boost Union theme does not store any personal data about any user.';
+
+// Additional layout setting tab title.
+$string['additionallayoutsettings'] = 'Additional Layout Settings';
+
+// Footnote settings.
+$string['footnoteheadingsetting'] = 'Footnote';
+$string['footnoteheadingsetting_desc'] = 'The following setting allows to add an additional region for displaying a footnote.';
+$string['footnotesetting'] = 'Footnote';
+$string['footnotesetting_desc'] = 'Whatever you add to this textarea will be displayed at the end of a page, in the footer on every page (not the floating footer) (for the layouts "drawers", "columns2" and "login"). Content in this area could be for example the copyright, the terms of use and the name of your organisation. <br/> If you want to remove the footnote again, just empty the text area.';
+