+++ CHANGES.md
@@ -6,6 +6,8 @@ Changes
 
 ### Unreleased
 
+=======
+* 2022-04-30 - Do write a footnote to the footer.
 * 2022-06-20 - Allow full Behat runs with Boost Campus suite, fixes #14.
 * 2022-06-20 - Prepare settings.php page, solves #2.
 * 2022-06-20 - Fill README.md, helps to resolve #3.